Extract product value creation in a dedicated service
