+++ Generator/ProductGenerator.php
@@ -6,19 +6,12 @@ use Akeneo\Component\Classification\Repository\CategoryRepositoryInterface;
 use Doctrine\Common\Persistence\ObjectRepository;
 use Faker;
 use Pim\Bundle\CatalogBundle\Entity\Family;
-use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
-use Pim\Bundle\CatalogBundle\Model\AttributeOptionInterface;
-use Pim\Bundle\CatalogBundle\Model\ChannelInterface;
-use Pim\Bundle\CatalogBundle\Model\CurrencyInterface;
 use Pim\Bundle\CatalogBundle\Model\FamilyInterface;
-use Pim\Bundle\CatalogBundle\Model\LocaleInterface;
 use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\CurrencyRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\FamilyRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\GroupRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Bundle\DataGeneratorBundle\Generator\Product\ProductValueBuilder;
 use Pim\Bundle\DataGeneratorBundle\VariantGroupDataProvider;
 use Symfony\Component\Console\Helper\ProgressHelper;
 
@@ -33,13 +26,8 @@ class ProductGenerator implements GeneratorInterface
 {
     const DEFAULT_FILENAME = 'products.csv';
     const IDENTIFIER_PREFIX = 'id-';
-    const METRIC_UNIT = 'unit';
 
     const CATEGORY_FIELD = 'categories';
-
-    const DEFAULT_NUMBER_MIN = '0';
-    const DEFAULT_NUMBER_MAX = '1000';
-    const DEFAULT_NB_DECIMALS = '4';
     const DEFAULT_DELIMITER = ',';
 
     /** @var string */
@@ -60,15 +48,6 @@ class ProductGenerator implements GeneratorInterface
     /** @var array */
     protected $attributesByFamily;
 
-    /** @var array */
-    protected $currencies;
-
-    /** @var array */
-    protected $channels;
-
-    /** @var array */
-    protected $locales;
-
     /** @var FamilyRepositoryInterface */
     protected $familyRepository;
 
@@ -77,15 +56,6 @@ class ProductGenerator implements GeneratorInterface
 
     /** @var string */
     protected $identifierCode;
-
-    /** @var LocaleRepositoryInterface */
-    protected $localeRepository;
-
-    /** @var ChannelRepositoryInterface */
-    protected $channelRepository;
-
-    /** @var CurrencyRepositoryInterface */
-    protected $currencyRepository;
 
     /** @var CategoryRepositoryInterface */
     protected $categoryRepository;
@@ -108,28 +78,25 @@ class ProductGenerator implements GeneratorInterface
     /** @var array */
     protected $headers;
 
+    /** @var ProductValueBuilder */
+    protected $valueBuilder;
+
     /**
+     * @param ProductValueBuilder          $valueBuilder
      * @param FamilyRepositoryInterface    $familyRepository
      * @param AttributeRepositoryInterface $attributeRepository
-     * @param ChannelRepositoryInterface   $channelRepository
-     * @param LocaleRepositoryInterface    $localeRepository
-     * @param CurrencyRepositoryInterface  $currencyRepository
      * @param CategoryRepositoryInterface  $categoryRepository
      * @param GroupRepositoryInterface     $groupRepository
      */
     public function __construct(
+        ProductValueBuilder $valueBuilder,
         FamilyRepositoryInterface $familyRepository,
         AttributeRepositoryInterface $attributeRepository,
-        ChannelRepositoryInterface $channelRepository,
-        LocaleRepositoryInterface $localeRepository,
-        CurrencyRepositoryInterface $currencyRepository,
         CategoryRepositoryInterface $categoryRepository,
         GroupRepositoryInterface $groupRepository
     ) {
+        $this->valueBuilder        = $valueBuilder;
         $this->familyRepository    = $familyRepository;
-        $this->channelRepository   = $channelRepository;
-        $this->localeRepository    = $localeRepository;
-        $this->currencyRepository  = $currencyRepository;
         $this->categoryRepository  = $categoryRepository;
         $this->attributeRepository = $attributeRepository;
         $this->groupRepository     = $groupRepository;
@@ -196,6 +163,7 @@ class ProductGenerator implements GeneratorInterface
         if (isset($globalConfig['seed'])) {
             $this->faker->seed($globalConfig['seed']);
         }
+        $this->valueBuilder->setFakerGenerator($this->faker);
 
         for ($i = $startIndex; $i < ($startIndex + $count); $i++) {
             $product = [];
@@ -204,7 +172,6 @@ class ProductGenerator implements GeneratorInterface
             $product['family'] = $family->getCode();
             $product['groups'] = '';
 
-            /** @var VariantGroupDataProvider $variantGroupDataProvider */
             $variantGroupDataProvider = $this->getNextVariantGroupProvider();
             $variantGroupAttributes = [];
 
@@ -254,263 +221,17 @@ class ProductGenerator implements GeneratorInterface
     }
 
     /**
-     * Generate a value in term of one or several entries in the product array
-     *
-     * @param AbstractAttribute $attribute
+     * @param AttributeInterface $attribute
      *
      * @return array
      */
-    protected function generateValue(AbstractAttribute $attribute)
+    protected function generateValue(AttributeInterface $attribute)
     {
-        $valueData = [];
-        $keys = $this->getAttributeKeys($attribute);
-
-        foreach ($keys as $key) {
-            $valueData[$key] = $this->generateValueData($attribute, $key);
-        }
-
-        return $valueData;
-    }
-
-    /**
-     * Provides the potential column keys for this attribute
-     *
-     * @param AbstractAttribute $attribute
-     *
-     * @return array
-     */
-    protected function getAttributeKeys(AbstractAttribute $attribute)
-    {
-        $keys = [];
-
-        $keys[$attribute->getCode()] = [];
-
-        $updatedKeys = [];
-        if ($attribute->isScopable() && $attribute->isLocalizable()) {
-            foreach ($this->getLocales() as $locale) {
-                foreach ($this->getChannels() as $channel) {
-                    foreach ($keys as $baseKey => $keyOptions) {
-                        $key = $baseKey.'-'.$locale->getCode().'-'.$channel->getCode();
-                        $updatedKeys[$key] = array_merge($keyOptions, ['locale' => $locale, 'channel' => $channel]);
-                    }
-                }
-            }
-            $keys = $updatedKeys;
-        } elseif ($attribute->isScopable() && !$attribute->isLocalizable()) {
-            foreach ($this->getChannels() as $channel) {
-                foreach ($keys as $baseKey => $keyOptions) {
-                    $key = $baseKey.'-'.$channel->getCode();
-                    $updatedKeys[$key] = array_merge($keyOptions, ['channel' => $channel]);
-                }
-            }
-            $keys = $updatedKeys;
-        } elseif (!$attribute->isScopable() && $attribute->isLocalizable()) {
-            foreach ($this->getLocales() as $locale) {
-                foreach ($keys as $baseKey => $keyOptions) {
-                    $key = $baseKey.'-'.$locale->getCode();
-                    $updatedKeys[$key] = array_merge($keyOptions, ['locale' => $locale]);
-                }
-            }
-            $keys = $updatedKeys;
-        }
-
-        switch ($attribute->getBackendType()) {
-            case 'prices':
-                $updatedKeys = [];
-
-                foreach ($keys as $key => $keyOptions) {
-                    foreach ($this->getCurrencies() as $currency) {
-                        $updatedKeys[$key.'-'.$currency->getCode()] = array_merge(
-                            $keyOptions,
-                            ['currency' => $currency]
-                        );
-                    }
-                }
-                $keys = $updatedKeys;
-                break;
-            case 'metric':
-                $updatedKeys = [];
-
-                foreach ($keys as $key => $keyOptions) {
-                    $updatedKeys[$key] = $keyOptions;
-                    $updatedKeys[$key.'-'.self::METRIC_UNIT] = $keyOptions;
-                }
-                $keys = $updatedKeys;
-                break;
-        }
-
-        $enabledKeys = [];
-        foreach ($keys as $key => $keyOptions) {
-            if ($this->isAttributeKeyValid($keyOptions)) {
-                $enabledKeys[] = $key;
-            }
-        }
-
-        return $enabledKeys;
-    }
-
-    /**
-     * Generate value content based on backend type
-     *
-     * @param AbstractAttribute $attribute
-     * @param string            $key
-     *
-     * @return string
-     */
-    protected function generateValueData(AbstractAttribute $attribute, $key)
-    {
-        $data = "";
-
         if (isset($this->forcedValues[$attribute->getCode()])) {
             return $this->forcedValues[$attribute->getCode()];
         }
 
-        if (preg_match('/-'.self::METRIC_UNIT.'$/', $key)) {
-            return $attribute->getDefaultMetricUnit();
-        }
-
-        switch ($attribute->getBackendType()) {
-            case "varchar":
-                $data = $this->generateVarcharData($attribute);
-                break;
-            case "text":
-                $data = $this->generateTextData();
-                break;
-            case "date":
-                $data = $this->generateDateData($attribute);
-                break;
-            case "metric":
-            case "decimal":
-            case "prices":
-                $data = $this->generateNumberData($attribute);
-                break;
-            case "boolean":
-                $data = $this->generateBooleanData();
-                break;
-            case "option":
-            case "options":
-                $data = $this->generateOptionData($attribute);
-                break;
-            default:
-                $data = '';
-                break;
-        }
-
-        return (string) $data;
-    }
-
-    /**
-     * Generate a varchar product value data
-     *
-     * @param AbstractAttribute $attribute
-     *
-     * @return string
-     */
-    protected function generateVarcharData(AbstractAttribute $attribute)
-    {
-        $validationRule = $attribute->getValidationRule();
-        switch ($validationRule) {
-            case 'url':
-                $varchar = $this->faker->url();
-                break;
-            default:
-                $varchar = $this->faker->sentence();
-                break;
-        }
-
-        return $varchar;
-    }
-
-    /**
-     * Generate a text product value data
-     *
-     * @return string
-     */
-    protected function generateTextData()
-    {
-        return $this->faker->sentence();
-    }
-
-    /**
-     * Generate a date product value data
-     *
-     * @param AbstractAttribute $attribute
-     *
-     * @return string
-     */
-    protected function generateDateData(AbstractAttribute $attribute)
-    {
-        $date = $this->faker->dateTimeBetween($attribute->getDateMin(), $attribute->getDateMax());
-        return $date->format('Y-m-d');
-    }
-
-    /**
-     * Generate number data
-     *
-     * @param AbstractAttribute $attribute
-     *
-     * @return string
-     */
-    protected function generateNumberData(AbstractAttribute $attribute)
-    {
-        $min = ($attribute->getNumberMin() != null) ? $attribute->getNumberMin() : self::DEFAULT_NUMBER_MIN;
-        $max = ($attribute->getNumberMax() != null) ? $attribute->getNumberMax() : self::DEFAULT_NUMBER_MAX;
-
-        $decimals = $attribute->isDecimalsAllowed() ? self::DEFAULT_NB_DECIMALS : 0;
-
-        $number = $this->faker->randomFloat($decimals, $min, $max);
-
-        return (string) $number;
-    }
-
-    /**
-     * Generate a boolean product value data
-     *
-     * @return string
-     */
-    protected function generateBooleanData()
-    {
-        return $this->faker->boolean() ? "1" : "0";
-    }
-
-    /**
-     * Generate option data
-     *
-     * @param AbstractAttribute $attribute
-     *
-     * @return string
-     */
-    protected function generateOptionData(AbstractAttribute $attribute)
-    {
-        $optionCode = "";
-
-        $option = $this->getRandomOptionFromAttribute($attribute);
-
-        if (is_object($option)) {
-            $optionCode = $option->getCode();
-        }
-
-        return $optionCode;
-    }
-
-    /**
-     * Get a random option from an attribute
-     *
-     * @param AbstractAttribute $attribute
-     *
-     * @return AttributeOptionInterface
-     */
-    protected function getRandomOptionFromAttribute(AbstractAttribute $attribute)
-    {
-        if (!isset($this->attributeOptions[$attribute->getCode()])) {
-            $this->attributeOptions[$attribute->getCode()] = [];
-
-            foreach ($attribute->getOptions() as $option) {
-                $this->attributeOptions[$attribute->getCode()][] = $option;
-            }
-        }
-
-        return $this->faker->randomElement($this->attributeOptions[$attribute->getCode()]);
+        return $this->valueBuilder->build($attribute);
     }
 
 
@@ -609,60 +330,6 @@ class ProductGenerator implements GeneratorInterface
     }
 
     /**
-     * Get all channels
-     *
-     * @return array
-     */
-    protected function getChannels()
-    {
-        if (null === $this->channels) {
-            $this->channels = [];
-            $channels = $this->channelRepository->findAll();
-            foreach ($channels as $channel) {
-                $this->channels[$channel->getCode()] = $channel;
-            }
-        }
-
-        return $this->channels;
-    }
-
-    /**
-     * Get active currencies
-     *
-     * @return array
-     */
-    protected function getCurrencies()
-    {
-        if (null === $this->currencies) {
-            $this->currencies = [];
-            $currencies = $this->currencyRepository->findBy(['activated' => 1]);
-            foreach ($currencies as $currency) {
-                $this->currencies[$currency->getCode()] = $currency;
-            }
-        }
-
-        return $this->currencies;
-    }
-
-    /**
-     * Get active locales
-     *
-     * @return array
-     */
-    protected function getLocales()
-    {
-        if (null === $this->locales) {
-            $this->locales = [];
-            $locales = $this->localeRepository->findBy(['activated' => 1]);
-            foreach ($locales as $locale) {
-                $this->locales[$locale->getCode()] = $locale;
-            }
-        }
-
-        return $this->locales;
-    }
-
-    /**
      * Get a random item from a repo
      *
      * @param Faker\Generator  $faker
@@ -715,68 +382,6 @@ class ProductGenerator implements GeneratorInterface
         $this->headers = array_unique(array_merge($this->headers, array_keys($product)));
 
         file_put_contents($this->tmpFile, serialize($product)."\n", FILE_APPEND);
-    }
-
-    /**
-     * Returns true if the key is valid, according to channels locale and currency.
-     *
-     * @param $options
-     *
-     * @return bool
-     */
-    private function isAttributeKeyValid(array $options)
-    {
-        if (isset($options['channel'])) {
-            $channel = $options['channel'];
-
-            if (isset($options['locale']) && !$this->hasChannelLocale($channel, $options['locale'])) {
-                return false;
-            }
-
-            if (isset($options['currency']) && !$this->hasChannelCurrency($channel, $options['currency'])) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Returns true if channel has activated locale.
-     *
-     * @param ChannelInterface $channel
-     * @param LocaleInterface  $locale
-     *
-     * @return bool
-     */
-    private function hasChannelLocale(ChannelInterface $channel, LocaleInterface $locale)
-    {
-        foreach ($channel->getLocaleCodes() as $availableLocale) {
-            if ($locale->getCode() === $availableLocale) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Returns true if channel has activated currency.
-     *
-     * @param ChannelInterface  $channel
-     * @param CurrencyInterface $currency
-     *
-     * @return bool
-     */
-    private function hasChannelCurrency(ChannelInterface $channel, CurrencyInterface $currency)
-    {
-        foreach ($channel->getCurrencies() as $availableCurrency) {
-            if ($currency->getCode() === $availableCurrency->getCode()) {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     /**
