+++ Resources/config/generators.yml
@@ -83,11 +83,9 @@ services:
     pim_data_generator.generator.product:
         class: %pim_data_generator.generator.product.class%
         arguments:
+            - '@pim_data_generator.product.value_builder'
             - '@pim_catalog.repository.family'
             - '@pim_catalog.repository.attribute'
-            - '@pim_catalog.repository.channel'
-            - '@pim_catalog.repository.locale'
-            - '@pim_catalog.repository.currency'
             - '@pim_catalog.repository.category'
             - '@pim_catalog.repository.group'
