+++ Resources/config/services.yml
@@ -1,0 +1,7 @@
+services:
+    pim_data_generator.product.value_builder:
+        class: 'Pim\Bundle\DataGeneratorBundle\Generator\Product\ProductValueBuilder'
+        arguments:
+            - '@pim_catalog.repository.channel'
+            - '@pim_catalog.repository.locale'
+            - '@pim_catalog.repository.currency'
